requirements.txt
openai  
python-dotenv
qdrant-client
streamlit
streamlit-audiorecorder 
pydub
pydantic
instructor
pandas
pathlib


page_functions